requirements.txt
chromadb==0.4.13
gradio==4.41.0
langchain==0.2.12
langchain_community==0.2.11
langchain_experimental==0.0.64
langchain_postgres==0.0.9
nltk~=3.9.1
pillow==11.0.0
psycopg2==2.9.9
PyMuPDF==1.24.13
sentence_transformers==3.0.1
setuptools==69.5.1
transformers==4.42.4
textwrap3==0.9.2
neo4j==5.26.0
python-dotenv~=1.0.1
pyneoinstance~=2.0.2
poppler-utils==0.1.0
pytesseract==0.3.13
unstructured[all-docs]==0.16.5
lxml==5.3.0
ipython~=8.29.0